requirements.txt
absl-py==1.4.0
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.2
cssselect==1.2.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.18.0
exceptiongroup==1.1.2
executing==2.0.1
fastapi==0.101.0
fastjsonschema==2.18.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.1
fqdn==1.5.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
httptools==0.6.0
hyperlink==21.0.0
idna==3.4
ijson==3.2.3
incremental==22.10.0
ipykernel==6.26.0
ipython==8.17.2
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
Protego==0.2.1
protobuf==4.23.4
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycurl==7.45.2
pydantic==1.10.12
pydantic_core==2.4.0
PyDispatcher==2.0.7
Pygments==2.16.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
rsa==4.9
Scrapy==2.9.0
seaborn==0.12.2
segment-anything==1.0
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
stack-data==0.6.3
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tldextract==3.4.4
tomli==2.0.1
tornado==6.3.3
traitlets==5.13.0
Twisted==22.10.0
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
w3lib==2.1.1
watchfiles==0.19.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
zope.interface==6.0
